Move adding data to advanced, move texts to translation files
