+++ core-ui/src/shared/components/CreateSecretForm/AdvancedForm.js
@@ -1,13 +1,37 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { FormFieldset, FormLabel } from 'fundamental-react';
+import { KeyValueForm } from 'react-shared';
 
 import { CreateModal } from 'shared/components/CreateModal/CreateModal';
 import { LabelsInput } from 'components/Lambdas/components';
 import { SimpleForm } from './SimpleForm';
+import { DecodeSecretSwitch } from './DecodeSecretSwitch';
 
 export function AdvancedForm({ secret, setSecret, isEncoded, setEncoded }) {
   const { t, i18n } = useTranslation();
+
+  const setData = data => {
+    setSecret({ ...secret, data });
+  };
+
+  const secretDataContent = (
+    <KeyValueForm
+      data={secret.data}
+      setData={setData}
+      setValid={() => {}}
+      customHeaderAction={(entries, setEntries) => (
+        <DecodeSecretSwitch
+          entries={entries}
+          setEntries={setEntries}
+          isEncoded={isEncoded}
+          setEncoded={setEncoded}
+          i18n={i18n}
+        />
+      )}
+      i18n={i18n}
+    />
+  );
 
   return (
     <>
@@ -33,6 +57,10 @@ export function AdvancedForm({ secret, setSecret, isEncoded, setEncoded }) {
             }
           />
         </FormFieldset>
+      </CreateModal.Section>
+      <CreateModal.Section>
+        <FormLabel>{t('secrets.create-modal.simple.data')}</FormLabel>
+        {secretDataContent}
       </CreateModal.Section>
     </>
   );
