+++ core-ui/src/shared/components/CreateSecretForm/DecodeSecretSwitch.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { useTranslation } from 'react-i18next';
 import { Button } from 'fundamental-react';
 import { useNotification } from 'react-shared';
+
 import { base64Decode, base64Encode } from 'shared/helpers';
 
 DecodeSecretSwitch.propTypes = {
@@ -19,8 +21,10 @@ export function DecodeSecretSwitch({
   setEntries,
   isEncoded,
   setEncoded,
+  i18n,
 }) {
   const notification = useNotification();
+  const { t } = useTranslation(null, { i18n });
 
   const onClick = () => {
     try {
@@ -30,10 +34,9 @@ export function DecodeSecretSwitch({
       setEncoded(!isEncoded);
     } catch (e) {
       notification.notifyError({
-        title: 'Failed to decode the Secret data',
-        content:
-          "Some of the secret data couldn't be decoded. Correct them and try again: " +
-          e.message,
+        content: t('secrets.create-modal.messages.decoding-failed', {
+          error: e.message,
+        }),
       });
     }
   };
@@ -44,7 +47,7 @@ export function DecodeSecretSwitch({
       glyph={isEncoded ? 'show' : 'hide'}
       onClick={onClick}
     >
-      {isEncoded ? 'Decode' : 'Encode'}
+      {isEncoded ? t('secrets.decode') : t('secrets.encode')}
     </Button>
   );
 }
