+++ core-ui/src/shared/components/CreateSecretForm/SimpleForm.js
@@ -2,34 +2,12 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { FormFieldset, FormLabel, FormInput } from 'fundamental-react';
 
-import { K8sNameInput, Tooltip, KeyValueForm } from 'react-shared';
+import { K8sNameInput, Tooltip } from 'react-shared';
 import { CreateModal } from 'shared/components/CreateModal/CreateModal';
 import { LabelsInput } from 'components/Lambdas/components';
-import { DecodeSecretSwitch } from './DecodeSecretSwitch';
 
-export function SimpleForm({ secret, setSecret, isEncoded, setEncoded }) {
+export function SimpleForm({ secret, setSecret }) {
   const { t, i18n } = useTranslation();
-
-  const setData = data => {
-    setSecret({ ...secret, data });
-  };
-
-  const secretDataContent = (
-    <KeyValueForm
-      data={secret.data}
-      setData={setData}
-      setValid={() => {}}
-      customHeaderAction={(entries, setEntries) => (
-        <DecodeSecretSwitch
-          entries={entries}
-          setEntries={setEntries}
-          isEncoded={isEncoded}
-          setEncoded={setEncoded}
-        />
-      )}
-      i18n={i18n}
-    />
-  );
 
   return (
     <>
@@ -81,11 +59,6 @@ export function SimpleForm({ secret, setSecret, isEncoded, setEncoded }) {
             }
           />
         </FormFieldset>
-      </CreateModal.Section>
-
-      <CreateModal.Section>
-        <FormLabel>{t('secrets.create-modal.simple.data')}</FormLabel>
-        {secretDataContent}
       </CreateModal.Section>
     </>
   );
